django_app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==41.0.3
Django==5.1.2
django-admin-sortable2==2.2.3
django-allauth==65.0.2
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-jazzmin==3.0.1
django-role-permissions==3.2.0
django-select2==8.2.1
et-xmlfile==1.1.0
idna==3.10
iniconfig==2.0.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
